cp-22 RestApi
cp-22 RestApi

+++ spring-boot-angular-15-crud-example-master/spring-boot-server/src/main/java/com/bezkoder/spring/jpa/h2/repository/ContactUsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface ContactUsRepository extends JpaRepository<ContactUs, Long> {
+public interface ContactUsRepository extends JpaRepository<ContactUs, Integer> {
 }
